# api-gateway/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MessageBox API Gateway",
    "description": "API gateway to interact with the MessageBox backend micro-services set.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Development server",
      "variables": {}
    }
  ],
...
  "paths": {
...
    "/Clients": {
      "get": {
        "tags": ["Client"],
        "summary": "Get all clients",
        "description": "Get all the clients.",
        "operationId": "GetAllClients",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All clients in the system",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientSchema"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "post": {
        "tags": ["Client"],
        "summary": "Register a new client and admin user",
        "description": "Register a new client and register the initial admin user.",
        "operationId": "CreateClientAndUser",
        "security": [],
        "requestBody": {
          "description": "Client and User information required to create and register a client.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "client": {
                    "type": "object",
                    "properties": {
                      "clientName": {
                        "type": "string"
                      }
                    }
                  },
                  "user": {
                    "type": "object",
                    "properties": {
                      "email": {
                        "type": "string",
                        "format": "email"
                      },
                      "firstName": {
                        "type": "string"
                      },
                      "lastName": {
                        "type": "string"
                      },
                      "password": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successfully registered a new client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "x-codegen-request-body-name": "body",
        "parameters": []
      }
    },
    "/Clients/Profile": {
      "get": {
        "tags": ["Client"],
        "summary": "Get the logged-in client's profile",
        "description": "Get the logged-in client's profile.",
        "operationId": "GetClientProfile",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched the client profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/Clients/Roles": {
      "get": {
        "tags": ["Client", "Role"],
        "summary": "Get all roles for logged-in client",
        "description": "Get all the available roles for the logged-in client.",
        "operationId": "GetAllRolesForClient",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All roles available to the client",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleSchema"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "post": {
        "tags": ["Client", "Role"],
        "summary": "Create a new role for the logged-in client",
        "description": "Create a new role for the logged-in client.",
        "operationId": "ClientCreateNewRole",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Data required to create a new user role.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rootName": {
                    "type": "string"
                  },
                  "viewName": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successfully create a new role for the client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "x-codegen-request-body-name": "body",
        "parameters": []
      }
    },
    "/Clients/Teams": {
      "get": {
        "tags": ["Client", "Team"],
        "summary": "Get all teams for logged-in client",
        "description": "Get all the available teams for the logged-in client.",
        "operationId": "GetAllTeamsForClient",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All teams available to the client",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamSchema"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "post": {
        "tags": ["Client", "Team"],
        "summary": "Create a new team for the logged-in client",
        "description": "Create a new team for the logged-in client.",
        "operationId": "ClientCreateNewTeam",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Data required to create a new user team.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rootName": {
                    "type": "string"
                  },
                  "teamName": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successfully create a new role for the client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "x-codegen-request-body-name": "body",
        "parameters": []
      }
    },
    "/Clients/Users": {
      "get": {
        "tags": ["Client", "User"],
        "summary": "Get all users for client",
        "description": "Get all the users for client.",
        "operationId": "GetAllUsers",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search from firstName, lastName and email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Change page based on pageSize",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Return users pageSize",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Sorting order of the Id",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          },
          {
            "name": "roleId",
            "in": "query",
            "description": "Specific available user role",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All users in the system",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSchema"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "post": {
        "tags": ["Client"],
        "summary": "Signed-in user to create new user for client",
        "description": "Signed-in user to create new user for client.",
        "operationId": "ClientCreateUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Personal data required to register a user.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successfully registered user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "x-codegen-request-body-name": "body",
        "parameters": []
      }
    },
...
    "/Users/Profile": {
      "get": {
        "tags": ["User"],
        "summary": "Get the logged-in user's details.",
        "description": "Get the logged-in user's details.",
        "operationId": "GetLoggedInUserProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logged-in user profile details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "put": {
        "tags": ["User"],
        "summary": "Update user the logged-in user profile",
        "description": "Update the logged-in user.",
        "operationId": "UpdateLoggedInUserProfile",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Updated user details",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "UnauthorizedError": {
        "description": "Access token is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Failed validation on the server",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/FieldValidationError"
            }
          }
        }
      }
    },
    "schemas": {
      "FieldValidationError": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "propertyPath": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "SortDirection": {
        "type": "string",
        "enum": ["ASC", "DESC"]
      },
...
      "UserSchema": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "isActive": {
            "type": "boolean",
            "default": false
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ClientSchema": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RoleSchema": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "rootName": {
            "type": "string"
          },
          "viewName": {
            "type": "string"
          },
          "isUserDeletable": {
            "type": "boolean",
            "default": false
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TeamSchema": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string",
            "format": "uuid"
          },
          "rootName": {
            "type": "string"
          },
          "teamName": {
            "type": "string"
          },
          "isUserDeletable": {
            "type": "boolean",
            "default": false
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": []
}